+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Examples.hs
@@ -4,11 +4,13 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
+-- | The example transactions in this module are not valid transactions. We
+-- don't care, we are only interested in serialisation, not validation.
 module Test.Cardano.Ledger.Allegra.Examples (
   ledgerExamples,
-  mkAllegraBasedExampleTx,
-  exampleAllegraBasedTxBody,
-  exampleAllegraBasedShelleyTxBody,
+  exampleAllegraBasedTx,
+  exampleAllegraToBabbageTx,
+  exampleAllegraToConwayTx,
 ) where
 
 import Cardano.Ledger.Allegra
@@ -30,10 +32,10 @@ import Lens.Micro
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples,
   exampleCoin,
-  exampleShelleyBasedShelleyTxBody,
-  exampleShelleyBasedTxBody,
+  exampleShelleyBasedTx,
+  exampleShelleyToBabbageTx,
+  exampleShelleyToConwayTx,
   mkKeyHash,
-  mkShelleyBasedExampleTx,
   mkShelleyBasedLedgerExamples,
  )
 
@@ -44,53 +46,71 @@ ledgerExamples =
         DelegateeNotRegisteredDELEG @AllegraEra (mkKeyHash 1)
     )
     exampleCoin
-    (mkAllegraBasedExampleTx $ exampleAllegraBasedShelleyTxBody exampleCoin)
+    exampleAllegraToBabbageTx
     NoGenesis
 
-mkAllegraBasedExampleTx ::
+-- Complete transaction which is compatible with any era starting with Allegra.
+-- This transaction forms the basis on which future era transactions will be
+-- at the very least based on.
+exampleAllegraBasedTx ::
   forall era.
   ( EraTx era
   , AllegraEraTxAuxData era
   , AllegraEraScript era
+  , AllegraEraTxBody era
   ) =>
-  TxBody TopTx era ->
   Tx TopTx era
-mkAllegraBasedExampleTx txBody =
-  mkShelleyBasedExampleTx @era txBody
+exampleAllegraBasedTx =
+  addAllegraBasedTxFeatures exampleShelleyBasedTx
+
+-- Complete Allegra transaction that is compatible until Babbage era
+-- ('ConwayEra' is not an instance of 'ShelleyEraTxBody').
+exampleAllegraToBabbageTx ::
+  forall era.
+  ( EraTx era
+  , AllegraEraTxAuxData era
+  , AllegraEraScript era
+  , AllegraEraTxBody era
+  , ShelleyEraTxBody era
+  ) =>
+  Tx TopTx era
+exampleAllegraToBabbageTx =
+  addAllegraBasedTxFeatures exampleShelleyToBabbageTx
+
+-- Complete Allegra transaction that is compatible until Conway era
+-- ('DijkstraEra' is not an instance of 'ShelleyEraTxCert').
+exampleAllegraToConwayTx ::
+  forall era.
+  ( EraTx era
+  , AllegraEraTxAuxData era
+  , AllegraEraScript era
+  , AllegraEraTxBody era
+  , ShelleyEraTxCert era
+  ) =>
+  Tx TopTx era
+exampleAllegraToConwayTx =
+  addAllegraBasedTxFeatures exampleShelleyToConwayTx
+
+addAllegraBasedTxFeatures ::
+  forall era.
+  ( EraTx era
+  , AllegraEraTxAuxData era
+  , AllegraEraScript era
+  , AllegraEraTxBody era
+  ) =>
+  Tx TopTx era ->
+  Tx TopTx era
+addAllegraBasedTxFeatures tx =
+  tx
     & witsTxL . scriptTxWitsL <>~ Map.singleton (hashScript script) script
     & auxDataTxL
       %~ fmap
         ( \auxData ->
             auxData & nativeScriptsTxAuxDataL <>~ StrictSeq.singleton exampleTimelock
         )
+    & bodyTxL . vldtTxBodyL .~ ValidityInterval (SJust (SlotNo 2)) (SJust (SlotNo 4))
   where
     script = fromNativeScript exampleTimelock
-
-exampleAllegraBasedShelleyTxBody ::
-  forall era.
-  ( AllegraEraTxBody era
-  , ShelleyEraTxBody era
-  ) =>
-  Value era ->
-  TxBody TopTx era
-exampleAllegraBasedShelleyTxBody value =
-  mkAllegraBasedExampleTxBody (exampleShelleyBasedShelleyTxBody value)
-
-exampleAllegraBasedTxBody ::
-  forall era.
-  AllegraEraTxBody era =>
-  Value era ->
-  TxBody TopTx era
-exampleAllegraBasedTxBody value =
-  mkAllegraBasedExampleTxBody (exampleShelleyBasedTxBody value)
-
-mkAllegraBasedExampleTxBody ::
-  forall era.
-  AllegraEraTxBody era =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkAllegraBasedExampleTxBody txBody =
-  txBody & vldtTxBodyL .~ ValidityInterval (SJust (SlotNo 2)) (SJust (SlotNo 4))
 
 exampleTimelock :: AllegraEraScript era => NativeScript era
 exampleTimelock =