+++ eras/dijkstra/impl/CHANGELOG.md
@@ -49,6 +49,9 @@
 ### testlib
 
 * Add `genNonEmptyAccountBalanceIntervals`
+* In `Test.Cardano.Ledger.Dijkstra.Examples`:
+  - Remove `mkDijkstraBasedExampleTx`, `mkDijkstraBasedExampleTxBody`
+  - Add `exampleDijkstraBasedTx`
 
 ## 0.2.0.0
 
@@ -175,7 +178,7 @@
 * Move `cddl-files` to `cddl/data`.
 * Add full `HuddleSpec`.
 
-### `testlib`
+### testlib
 
 * Add `Test.Cardano.Ledger.Dijkstra.Imp.LedgerSpec`
 * Add `Test.Cardano.Ledger.Dijkstra.Imp.UtxoSpec`